third_party/blink/web_tests/external/wpt/editing/run/delete_3001-4000-expected.txt

This is a testharness.js-based test.
Found 27 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["delete",""]] "<p style=color:blue>foo<p style=color:rgba(0,0,255,1)>[]bar" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["delete",""]] "<p>foo<p style=color:brown>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<font color=\\"#a52a2a\\">bar</font></p>" but got "<p>foo<span style=\\"color:rgb(165, 42, 42)\\">bar</span></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["delete",""]] "<p>foo<p style=color:brown>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<font color=\\"#a52a2a\\">bar</font></p>" but got "<p>foo<span style=\\"color:rgb(165, 42, 42)\\">bar</span></p>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "<p><font color=blue>foo</font><p><font color=brown>[]bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font color=\\"blue\\">foo</font><font color=\\"brown\\">bar</font></p>" but got "<p><font color=\\"blue\\">foo</font><span style=\\"color:rgb(165, 42, 42)\\">bar</span></p>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "<p><font color=blue>foo</font><p><font color=brown>[]bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font color=\\"blue\\">foo</font><font color=\\"brown\\">bar</font></p>" but got "<p><font color=\\"blue\\">foo</font><span style=\\"color:rgb(165, 42, 42)\\">bar</span></p>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "<p>foo<p><font color=brown>[]bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<font color=\\"brown\\">bar</font></p>" but got "<p>foo<span style=\\"color:rgb(165, 42, 42)\\">bar</span></p>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "<p>foo<p><font color=brown>[]bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<font color=\\"brown\\">bar</font></p>" but got "<p>foo<span style=\\"color:rgb(165, 42, 42)\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["delete",""]] "<p style=background-color:aqua>foo<p style=background-color:tan>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"background-color:rgb(0, 255, 255)\\">foobar</p>" but got "<p style=\\"background-color:rgb(0, 255, 255)\\">foo<span style=\\"background-color:rgb(210, 180, 140)\\">bar</span></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["delete",""]] "<p style=background-color:aqua>foo<p style=background-color:tan>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"background-color:rgb(0, 255, 255)\\">foobar</p>" but got "<p style=\\"background-color:rgb(0, 255, 255)\\">foo<span style=\\"background-color:rgb(210, 180, 140)\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["delete",""]] "<p style=background-color:aqua>foo<p style=background-color:tan>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"background-color:rgb(0, 255, 255)\\">foobar</p>" but got "<p style=\\"background-color:rgb(0, 255, 255)\\">foo<span style=\\"background-color:rgb(210, 180, 140)\\">bar</span></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["delete",""]] "<p style=background-color:aqua>foo<p style=background-color:tan>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"background-color:rgb(0, 255, 255)\\">foobar</p>" but got "<p style=\\"background-color:rgb(0, 255, 255)\\">foo<span style=\\"background-color:rgb(210, 180, 140)\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["delete",""]] "<p>foo<p style=background-color:tan>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foobar</p>" but got "<p>foo<span style=\\"background-color:rgb(210, 180, 140)\\">bar</span></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["delete",""]] "<p>foo<p style=background-color:tan>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foobar</p>" but got "<p>foo<span style=\\"background-color:rgb(210, 180, 140)\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["delete",""]] "<p>foo<p style=background-color:tan>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foobar</p>" but got "<p>foo<span style=\\"background-color:rgb(210, 180, 140)\\">bar</span></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["delete",""]] "<p>foo<p style=background-color:tan>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foobar</p>" but got "<p>foo<span style=\\"background-color:rgb(210, 180, 140)\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["delete",""]] "<p style=text-decoration:underline>foo<p>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"text-decoration:underline\\">foo</span>bar</p>" but got "<p style=\\"text-decoration:underline\\">foobar</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["delete",""]] "<p style=text-decoration:underline>foo<p>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><u>foo</u>bar</p>" but got "<p style=\\"text-decoration:underline\\">foobar</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["delete",""]] "<p style=text-decoration:underline>foo<p>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"text-decoration:underline\\">foo</span>bar</p>" but got "<p style=\\"text-decoration:underline\\">foobar</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["delete",""]] "<p style=text-decoration:underline>foo<p>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><u>foo</u>bar</p>" but got "<p style=\\"text-decoration:underline\\">foobar</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["delete",""]] "<p style=text-decoration:underline>foo<p style=text-decoration:line-through>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"text-decoration:underline\\">foo</span><span style=\\"text-decoration:line-through\\">bar</span></p>" but got "<p style=\\"text-decoration:underline\\">foo<span style=\\"text-decoration-line:line-through\\">bar</span></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["delete",""]] "<p style=text-decoration:underline>foo<p style=text-decoration:line-through>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><u>foo</u><s>bar</s></p>" but got "<p style=\\"text-decoration:underline\\">foo<span style=\\"text-decoration-line:line-through\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["delete",""]] "<p style=text-decoration:underline>foo<p style=text-decoration:line-through>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"text-decoration:underline\\">foo</span><span style=\\"text-decoration:line-through\\">bar</span></p>" but got "<p style=\\"text-decoration:underline\\">foo<span style=\\"text-decoration-line:line-through\\">bar</span></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["delete",""]] "<p style=text-decoration:underline>foo<p style=text-decoration:line-through>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><u>foo</u><s>bar</s></p>" but got "<p style=\\"text-decoration:underline\\">foo<span style=\\"text-decoration-line:line-through\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["delete",""]] "<p>foo<p style=text-decoration:line-through>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<span style=\\"text-decoration:line-through\\">bar</span></p>" but got "<p>foo<span style=\\"text-decoration-line:line-through\\">bar</span></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["delete",""]] "<p>foo<p style=text-decoration:line-through>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<s>bar</s></p>" but got "<p>foo<span style=\\"text-decoration-line:line-through\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["delete",""]] "<p>foo<p style=text-decoration:line-through>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<span style=\\"text-decoration:line-through\\">bar</span></p>" but got "<p>foo<span style=\\"text-decoration-line:line-through\\">bar</span></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["delete",""]] "<p>foo<p style=text-decoration:line-through>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<s>bar</s></p>" but got "<p>foo<span style=\\"text-decoration-line:line-through\\">bar</span></p>"
Harness: the test ran to completion.
